Add a test for traceback formatting behavior
The assertions in the test target several milestones:

* Showing the filename (currently it shows `"<string>"`)
* Showing standardized path separators for the sqlite-based file
* Showing the source code of the line (currently it is blank)
