+++ tests/installed-projects/filesystem/package_filesystem-1.1.1.dist-info/RECORD
@@ -7,4 +7,5 @@ package_filesystem-1.1.1.dist-info/direct_url.json,,
 package_filesystem/__init__.py,sha256=47DEQpj8HBSa-_TImW-5JCeuQeRkm5NMpJWZG3hSuFU,0
 package_filesystem/resource.txt,sha256=UzUTwTl8uMzsBYUrUlFL7NX9jJwhUJ97wvXUYMYUPdg,9
 package_filesystem/shift_jis.py,sha256=s5AYdLUzNAUliEPKSpJzWaVSJ8MopnN1YtP_gBXOFqs,31
+package_filesystem/zero_division.py,sha256=5fit7pGNdw81C5Iz8GezFqJKjOpB8Flg1qSKY10ZwqY,52
 package_filesystem/あ.py,sha256=hudcY6c59JQOUKDM1e7HT-ZDoMzaQ38C40TaJgBLcBA,29
