+++ tests/installed-projects/filesystem/package_filesystem-1.1.1.dist-info/direct_url.json
@@ -1,1 +1,1 @@
-{"archive_info": {"hash": "sha256=1e6afe77f443987349ad0af6c8a32a54ad13fbce8f1841bae659a420a202c908", "hashes": {"sha256": "1e6afe77f443987349ad0af6c8a32a54ad13fbce8f1841bae659a420a202c908"}}, "url": "file:///package_filesystem-1.1.1-py3-none-any.whl"}
+{"archive_info": {"hash": "sha256=bc344b40df21336094949388432b8357903ba00f83d68ac33a725d357c24f7c6", "hashes": {"sha256": "bc344b40df21336094949388432b8357903ba00f83d68ac33a725d357c24f7c6"}}, "url": "file:///package_filesystem-1.1.1-py3-none-any.whl"}
