+++ tests/installed-projects/sqlite/package_sqlite-2.2.2.dist-info/RECORD
@@ -7,4 +7,5 @@ package_sqlite-2.2.2.dist-info/direct_url.json,,
 package_sqlite/__init__.py,sha256=47DEQpj8HBSa-_TImW-5JCeuQeRkm5NMpJWZG3hSuFU,0
 package_sqlite/resource.txt,sha256=UzUTwTl8uMzsBYUrUlFL7NX9jJwhUJ97wvXUYMYUPdg,9
 package_sqlite/shift_jis.py,sha256=s5AYdLUzNAUliEPKSpJzWaVSJ8MopnN1YtP_gBXOFqs,31
+package_sqlite/zero_division.py,sha256=5fit7pGNdw81C5Iz8GezFqJKjOpB8Flg1qSKY10ZwqY,52
 package_sqlite/あ.py,sha256=hudcY6c59JQOUKDM1e7HT-ZDoMzaQ38C40TaJgBLcBA,29
