+++ tests/installed-projects/sqlite/package_sqlite-2.2.2.dist-info/direct_url.json
@@ -1,1 +1,1 @@
-{"archive_info": {"hash": "sha256=249b16071d2f84da12796af3690082d9b07a5dbeddeaf47d07bca82aa92b1959", "hashes": {"sha256": "249b16071d2f84da12796af3690082d9b07a5dbeddeaf47d07bca82aa92b1959"}}, "url": "file:///package_sqlite-2.2.2-py3-none-any.whl"}
+{"archive_info": {"hash": "sha256=e079cc91a99604a41d28f76b6ebf205f0678be5e5f30d64a6f554bab401a730c", "hashes": {"sha256": "e079cc91a99604a41d28f76b6ebf205f0678be5e5f30d64a6f554bab401a730c"}}, "url": "file:///package_sqlite-2.2.2-py3-none-any.whl"}
