+++ tests/source-projects/package/filesystem/src/package_filesystem/zero_division.py
@@ -1,0 +1,2 @@
+def trigger_zero_division_error():
+    print(1 / 0)
