+++ tests/source-projects/package/sqlite/src/package_sqlite/zero_division.py
@@ -1,0 +1,2 @@
+def trigger_zero_division_error():
+    print(1 / 0)
