+++ tests/test_importing.py
@@ -2,8 +2,6 @@ import contextlib
 import importlib.machinery
 import importlib.metadata
 import importlib.resources
-import pathlib
-import sqlite3
 import sys
 import types
 import uuid
@@ -13,20 +11,6 @@ import pytest
 import sqliteimport
 import sqliteimport.accessor
 import sqliteimport.bundler
-
-installed_projects = pathlib.Path(__file__).parent / "installed-projects"
-sys.path.append(str(installed_projects / "filesystem"))
-
-
-@pytest.fixture(scope="module")
-def database():
-    with sqlite3.connect(":memory:") as connection:
-        accessor = sqliteimport.accessor.Accessor(connection)
-        accessor.initialize_database()
-        sqliteimport.bundler.bundle(installed_projects / "sqlite", accessor)
-        sqliteimport.load(connection)
-
-        yield connection
 
 
 @pytest.mark.parametrize(
